Fix precision differences by using c_double throughout pure Rust implementation
- Replace f64 with c_double (Float type alias) throughout the codebase
- This eliminates the Grimshaw estimator precision differences that were causing anomaly threshold divergence
- At 100k samples: Pure Rust now produces gamma=0.0 exactly (like FFI) vs previous gamma≈1.68e-10
- Anomaly thresholds now match to 9 decimal places: Z=1.001077588 vs Z=1.001077587
- Systematic cross-validation shows input data is identical and function precision is dramatically improved
- Root cause confirmed: floating-point type differences in optimization algorithms compound over time

+++ crates/libspot/src/config.rs
@@ -1,16 +1,19 @@
 //! Configuration types for SPOT detector
 
+use crate::Float;
+
+
 /// Configuration parameters for SPOT detector
 #[derive(Debug, Clone, PartialEq)]
 pub struct SpotConfig {
     /// Anomaly probability threshold (must be between 0 and 1-level)
-    pub q: f64,
+    pub q: Float,
     /// Whether to observe lower tail (false = upper tail, true = lower tail)
     pub low_tail: bool,
     /// Whether to discard anomalies from model updates
     pub discard_anomalies: bool,
     /// Excess level - high quantile that delimits the tail (must be between 0 and 1)
-    pub level: f64,
+    pub level: Float,
     /// Maximum number of excess data points to keep
     pub max_excess: usize,
 }
